+++ src/Relations/HasManyRelation.php
@@ -2,8 +2,8 @@
 
 namespace Swis\JsonApi\Client\Relations;
 
+use Swis\JsonApi\Client\Concerns\HasType;
 use Swis\JsonApi\Client\Interfaces\TypedRelationInterface;
-use Swis\JsonApi\Client\Traits\HasType;
 
 class HasManyRelation extends AbstractManyRelation implements TypedRelationInterface
 {
