+++ src/Relations/HasOneRelation.php
@@ -2,8 +2,8 @@
 
 namespace Swis\JsonApi\Client\Relations;
 
+use Swis\JsonApi\Client\Concerns\HasType;
 use Swis\JsonApi\Client\Interfaces\TypedRelationInterface;
-use Swis\JsonApi\Client\Traits\HasType;
 
 class HasOneRelation extends AbstractOneRelation implements TypedRelationInterface
 {
